requirements.txt
Pillow
redis
flask
deepface
werkzeug